refactor: change structure KeysArray. Merged all scripts in one file (eslint error: import/extensions when import JS-scripts)
