--- css/style.scss
@@ -1,185 +1,0 @@
-$sizePage: 100%;
-@mixin wrapper {
-  position: absolute;
-  display: flex;
-  left: calc(50% - #{$sizePage/2});
-  width: $sizePage;
-  max-width: $sizePage;
-}
-.container {
-  @include wrapper;
-  height: 100%;
-  bottom: 0;
-
-  font-family: Arial, Helvetica, sans-serif;
-  font-size: 16px;
-  font-weight: 400;
-  justify-content: center;
-  .helper {
-    position: absolute;
-    display: block;
-    width: 770px;
-    height: 100px;
-    align-items: center;
-    justify-self: center;
-    text-align: center;
-    kbd {
-      border: rgba(0, 20, 255, 0.22) solid 1px;
-      border-radius: 3px;
-      background-color: #dbdbdb;
-    }
-    p {
-      margin: 0;
-    }
-    h1 {
-      margin: 10px;
-    }
-  }
-  .output {
-    position: absolute;
-    display: flex;
-    top: 100px;
-    width: 770px;
-    height: 150px;
-    text-align: center;
-    justify-content: center;
-    align-items: center;
-    textarea {
-      width: 100%;
-      height: 100%;
-      padding: 12px 20px;
-      box-sizing: border-box;
-      border: 2px solid #ccc;
-      border-radius: 4px;
-      background-color: #f8f8f8;
-      resize: none;
-      font-size: 18px;
-      font-family: "Ubuntu Mono";
-    }
-  }
-  .keyboard {
-    position: absolute;
-    display: grid;
-    grid-template-rows: repeat(5, 40px);
-    grid-template-columns: 100%;
-    row-gap: 10px;
-    height: 240px;
-    width: 750px;
-    top: 280px;
-    justify-content: center;
-    padding: 10px;
-    border: #949393 solid 1px;
-    background-color: #d4d1d1;
-    border-radius: 5px;
-  }
-  .keyboard-row {
-    display: grid;
-    grid-template-rows: 40px;
-    column-gap: 10px;
-    justify-content: center;
-    align-items: center;
-    text-align: center;
-  }
-  [data-row="0"] {
-    grid-template-columns: repeat(13, 40px) 100px;
-  }
-  [data-row="1"] {
-    grid-template-columns: 40px repeat(13, 40px) 50px;
-  }
-  [data-row="2"] {
-    grid-template-columns: 90px repeat(11, 40px) 100px;
-  }
-  [data-row="3"] {
-    grid-template-columns: 90px repeat(11, 40px) 100px;
-  }
-  [data-row="4"] {
-    grid-template-columns: repeat(3, 40px) 340px repeat(4, 40px) 50px;
-  }
-
-  .key {
-    display: flex;
-    cursor: pointer;
-    flex-direction: column;
-    text-align: center;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-    height: 100%;
-    .letter {
-      text-align: center;
-      width: 100%;
-      img {
-        width: 20px;
-        height: 20px;
-      }
-      .icon {
-        background-image: url('../assets/images/Language.svg');
-        width: 20px;
-        height: 20px;
-        background-size: 20px;
-      }
-    }
-    .sub-letter {
-      text-align: center;
-      width: 100%;
-      .icon {
-        background-image: url('../assets/images/Language.svg');
-        width: 20px;
-        height: 20px;
-        background-size: 20px;
-      }
-    }
-    [data-display="true"] {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-    }
-    [data-display="false"] {
-      display: none;
-    }
-  }
-  .key:hover {
-    background-color: #c3b8f0;
-    border-color: #5e3aec;
-    box-shadow: 0 2px 5px rgba(81, 42, 236, 0.5), 0 1px 1px rgba(81, 42, 236, 0.45);
-    font-weight: 600;
-  }
-  [data-class="special"] {
-    background-color: #bbbbbb;
-    border: #959595 solid 1px;
-    border-radius: 5px;
-    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.25), 0 1px 1px rgba(0, 0, 0, 0.22);
-  }
-  [data-class="press"] {
-    background-color: #c3b8f0;
-    border-radius: 20px;
-    border: #5e3aec solid 1px;
-    box-shadow: 0 2px 5px rgba(81, 42, 236, 0.5), 0 1px 1px rgba(81, 42, 236, 0.45);
-    font-weight: 600;
-
-    animation: press 0.25s;
-  }
-  [data-class="unpress"] {
-    border: #959595 solid 1px;
-    background-color: #dfdfdf;
-    border-radius: 5px;
-    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.25), 0 1px 1px rgba(0, 0, 0, 0.22);
-  }
-  @keyframes press {
-    0% {
-      opacity: 1;
-      border: #959595 solid 1px;
-      background-color: #dfdfdf;
-      border-radius: 5px;
-      box-shadow: 0 2px 5px rgba(0, 0, 0, 0.25), 0 1px 1px rgba(0, 0, 0, 0.22);
-    }
-    100% {
-      opacity: 1;
-      background-color: #c3b8f0;
-      border-radius: 20px;
-      border: #5e3aec solid 1px;
-      box-shadow: 0 2px 5px rgba(81, 42, 236, 0.5), 0 1px 1px rgba(81, 42, 236, 0.45);
-      font-weight: 600;
-    }
-  }
-}
