--- js/en.js
@@ -1,327 +1,0 @@
-export default [
-  {
-    Letter: '`',
-    LetterShift: '~',
-    code: 'Backquote',
-  },
-  {
-    Letter: '1',
-    LetterShift: '!',
-    code: 'Digit1',
-  },
-  {
-    Letter: '2',
-    LetterShift: '@',
-    code: 'Digit2',
-  },
-  {
-    Letter: '3',
-    LetterShift: '#',
-    code: 'Digit3',
-  },
-  {
-    Letter: '4',
-    LetterShift: '$',
-    code: 'Digit4',
-  },
-  {
-    Letter: '5',
-    LetterShift: '%',
-    code: 'Digit5',
-  },
-  {
-    Letter: '6',
-    LetterShift: '^',
-    code: 'Digit6',
-  },
-  {
-    Letter: '7',
-    LetterShift: '&',
-    code: 'Digit7',
-  },
-  {
-    Letter: '8',
-    LetterShift: '*',
-    code: 'Digit8',
-  },
-  {
-    Letter: '9',
-    LetterShift: '(',
-    code: 'Digit9',
-  },
-  {
-    Letter: '0',
-    LetterShift: ')',
-    code: 'Digit0',
-  },
-  {
-    Letter: '-',
-    LetterShift: '_',
-    code: 'Minus',
-  },
-  {
-    Letter: '=',
-    LetterShift: '+',
-    code: 'Equal',
-  },
-  {
-    Letter: 'Backspace',
-    LetterShift: 'Backspace',
-    code: 'Backspace',
-  },
-  {
-    Letter: 'Delete',
-    LetterShift: 'Delete',
-    code: 'Delete',
-  },
-  {
-    Letter: 'Tab',
-    LetterShift: 'Tab',
-    code: 'Tab',
-  },
-  {
-    Letter: 'q',
-    LetterShift: 'Q',
-    code: 'KeyQ',
-  },
-  {
-    Letter: 'w',
-    LetterShift: 'W',
-    code: 'KeyW',
-  },
-  {
-    Letter: 'e',
-    LetterShift: 'E',
-    code: 'KeyE',
-  },
-  {
-    Letter: 'r',
-    LetterShift: 'R',
-    code: 'KeyR',
-  },
-  {
-    Letter: 't',
-    LetterShift: 'T',
-    code: 'KeyT',
-  },
-  {
-    Letter: 'y',
-    LetterShift: 'Y',
-    code: 'KeyY',
-  },
-  {
-    Letter: 'u',
-    LetterShift: 'U',
-    code: 'KeyU',
-  },
-  {
-    Letter: 'i',
-    LetterShift: 'I',
-    code: 'KeyI',
-  },
-  {
-    Letter: 'o',
-    LetterShift: 'O',
-    code: 'KeyO',
-  },
-  {
-    Letter: 'p',
-    LetterShift: 'P',
-    code: 'KeyP',
-  },
-  {
-    Letter: '[',
-    LetterShift: '{',
-    code: 'BracketLeft',
-  },
-  {
-    Letter: ']',
-    LetterShift: '}',
-    code: 'BracketRight',
-  },
-  {
-    Letter: 'Enter',
-    LetterShift: 'Enter',
-    code: 'Enter',
-  },
-  {
-    Letter: 'CapsLock',
-    LetterShift: 'CapsLock',
-    code: 'CapsLock',
-  },
-  {
-    Letter: 'a',
-    LetterShift: 'A',
-    code: 'KeyA',
-  },
-  {
-    Letter: 's',
-    LetterShift: 'S',
-    code: 'KeyS',
-  },
-  {
-    Letter: 'd',
-    LetterShift: 'D',
-    code: 'KeyD',
-  },
-  {
-    Letter: 'f',
-    LetterShift: 'F',
-    code: 'KeyF',
-  },
-  {
-    Letter: 'g',
-    LetterShift: 'G',
-    code: 'KeyG',
-  },
-  {
-    Letter: 'h',
-    LetterShift: 'H',
-    code: 'KeyH',
-  },
-  {
-    Letter: 'j',
-    LetterShift: 'J',
-    code: 'KeyJ',
-  },
-  {
-    Letter: 'k',
-    LetterShift: 'K',
-    code: 'KeyK',
-  },
-  {
-    Letter: 'l',
-    LetterShift: 'L',
-    code: 'KeyL',
-  },
-  {
-    Letter: ';',
-    LetterShift: ':',
-    code: 'Semicolon',
-  },
-  {
-    Letter: "'",
-    LetterShift: '"',
-    code: 'Quote',
-  },
-  {
-    Letter: '\\',
-    LetterShift: '|',
-    code: 'Backslash',
-  },
-  {
-    Letter: 'Shift',
-    LetterShift: 'Shift',
-    code: 'ShiftLeft',
-  },
-  {
-    Letter: '<',
-    LetterShift: '>',
-    code: 'IntlBackslash',
-  },
-  {
-    Letter: 'z',
-    LetterShift: 'Z',
-    code: 'KeyZ',
-  },
-  {
-    Letter: 'x',
-    LetterShift: 'X',
-    code: 'KeyX',
-  },
-  {
-    Letter: 'c',
-    LetterShift: 'C',
-    code: 'KeyC',
-  },
-  {
-    Letter: 'v',
-    LetterShift: 'V',
-    code: 'KeyV',
-  },
-  {
-    Letter: 'b',
-    LetterShift: 'B',
-    code: 'KeyB',
-  },
-  {
-    Letter: 'n',
-    LetterShift: 'N',
-    code: 'KeyN',
-  },
-  {
-    Letter: 'm',
-    LetterShift: 'M',
-    code: 'KeyM',
-  },
-  {
-    Letter: ',',
-    LetterShift: '<',
-    code: 'Comma',
-  },
-  {
-    Letter: '.',
-    LetterShift: '>',
-    code: 'Period',
-  },
-  {
-    Letter: '/',
-    LetterShift: '?',
-    code: 'Slash',
-  },
-  {
-    Letter: 'Shift',
-    LetterShift: 'Shift',
-    code: 'ShiftRight',
-  },
-  {
-    Letter: 'Ctrl',
-    LetterShift: 'Ctrl',
-    code: 'ControlLeft',
-  },
-  {
-    Letter: 'Alt',
-    LetterShift: 'Alt',
-    code: 'AltLeft',
-  },
-  {
-    Letter: ' ',
-    LetterShift: ' ',
-    code: 'Space',
-  },
-  {
-    Letter: 'Alt',
-    LetterShift: 'Alt',
-    code: 'AltRight',
-  },
-  {
-    Letter: 'Ctrl',
-    LetterShift: 'Ctrl',
-    code: 'ControlRight',
-  },
-  {
-    Letter: '&larr;',
-    LetterShift: '&larr;',
-    code: 'ArrowLeft',
-  },
-  {
-    Letter: '&uarr;',
-    LetterShift: '&uarr;',
-    code: 'ArrowUp',
-  },
-  {
-    Letter: '&darr;',
-    LetterShift: '&darr;',
-    code: 'ArrowDown',
-  },
-  {
-    Letter: '&rarr;',
-    LetterShift: '&rarr;',
-    code: 'ArrowRight',
-  },
-  {
-    Letter: '<div class="icon"></div>',
-    LetterShift: '<div class="icon"></div>',
-    code: 'LNG',
-  },
-];
