--- js/ru.js
@@ -1,327 +1,0 @@
-export default [
-  {
-    Letter: 'CapsLock',
-    LetterShift: 'CapsLock',
-    code: 'CapsLock',
-  },
-  {
-    Letter: 'ё',
-    LetterShift: 'Ё',
-    code: 'Backquote',
-  },
-  {
-    Letter: '1',
-    LetterShift: '!',
-    code: 'Digit1',
-  },
-  {
-    Letter: '2',
-    LetterShift: '"',
-    code: 'Digit2',
-  },
-  {
-    Letter: '3',
-    LetterShift: '№',
-    code: 'Digit3',
-  },
-  {
-    Letter: '4',
-    LetterShift: ';',
-    code: 'Digit4',
-  },
-  {
-    Letter: '5',
-    LetterShift: '%',
-    code: 'Digit5',
-  },
-  {
-    Letter: '6',
-    LetterShift: ':',
-    code: 'Digit6',
-  },
-  {
-    Letter: '7',
-    LetterShift: '?',
-    code: 'Digit7',
-  },
-  {
-    Letter: '8',
-    LetterShift: '*',
-    code: 'Digit8',
-  },
-  {
-    Letter: '9',
-    LetterShift: '(',
-    code: 'Digit9',
-  },
-  {
-    Letter: '0',
-    LetterShift: ')',
-    code: 'Digit0',
-  },
-  {
-    Letter: '-',
-    LetterShift: '_',
-    code: 'Minus',
-  },
-  {
-    Letter: '=',
-    LetterShift: '+',
-    code: 'Equal',
-  },
-  {
-    Letter: 'Backspace',
-    LetterShift: 'Backspace',
-    code: 'Backspace',
-  },
-  {
-    Letter: 'Delete',
-    LetterShift: 'Delete',
-    code: 'Delete',
-  },
-  {
-    Letter: 'Tab',
-    LetterShift: 'Tab',
-    code: 'Tab',
-  },
-  {
-    Letter: 'й',
-    LetterShift: 'Й',
-    code: 'KeyQ',
-  },
-  {
-    Letter: 'ц',
-    LetterShift: 'Ц',
-    code: 'KeyW',
-  },
-  {
-    Letter: 'у',
-    LetterShift: 'У',
-    code: 'KeyE',
-  },
-  {
-    Letter: 'к',
-    LetterShift: 'К',
-    code: 'KeyR',
-  },
-  {
-    Letter: 'е',
-    LetterShift: 'Е',
-    code: 'KeyT',
-  },
-  {
-    Letter: 'н',
-    LetterShift: 'Н',
-    code: 'KeyY',
-  },
-  {
-    Letter: 'г',
-    LetterShift: 'Г',
-    code: 'KeyU',
-  },
-  {
-    Letter: 'ш',
-    LetterShift: 'Ш',
-    code: 'KeyI',
-  },
-  {
-    Letter: 'щ',
-    LetterShift: 'Щ',
-    code: 'KeyO',
-  },
-  {
-    Letter: 'з',
-    LetterShift: 'З',
-    code: 'KeyP',
-  },
-  {
-    Letter: 'х',
-    LetterShift: 'Х',
-    code: 'BracketLeft',
-  },
-  {
-    Letter: 'ъ',
-    LetterShift: 'Ъ',
-    code: 'BracketRight',
-  },
-  {
-    Letter: 'Enter',
-    LetterShift: 'Enter',
-    code: 'Enter',
-  },
-  {
-    Letter: 'ф',
-    LetterShift: 'Ф',
-    code: 'KeyA',
-  },
-  {
-    Letter: 'ы',
-    LetterShift: 'Ы',
-    code: 'KeyS',
-  },
-  {
-    Letter: 'в',
-    LetterShift: 'В',
-    code: 'KeyD',
-  },
-  {
-    Letter: 'а',
-    LetterShift: 'А',
-    code: 'KeyF',
-  },
-  {
-    Letter: 'п',
-    LetterShift: 'П',
-    code: 'KeyG',
-  },
-  {
-    Letter: 'р',
-    LetterShift: 'Р',
-    code: 'KeyH',
-  },
-  {
-    Letter: 'о',
-    LetterShift: 'О',
-    code: 'KeyJ',
-  },
-  {
-    Letter: 'л',
-    LetterShift: 'Л',
-    code: 'KeyK',
-  },
-  {
-    Letter: 'д',
-    LetterShift: 'Д',
-    code: 'KeyL',
-  },
-  {
-    Letter: 'ж',
-    LetterShift: 'Ж',
-    code: 'Semicolon',
-  },
-  {
-    Letter: 'э',
-    LetterShift: 'Э',
-    code: 'Quote',
-  },
-  {
-    Letter: '\\',
-    LetterShift: '/',
-    code: 'Backslash',
-  },
-  {
-    Letter: 'Shift',
-    LetterShift: 'Shift',
-    code: 'ShiftLeft',
-  },
-  {
-    Letter: '/',
-    LetterShift: '|',
-    code: 'IntlBackslash',
-  },
-  {
-    Letter: 'я',
-    LetterShift: 'Я',
-    code: 'KeyZ',
-  },
-  {
-    Letter: 'ч',
-    LetterShift: 'Ч',
-    code: 'KeyX',
-  },
-  {
-    Letter: 'с',
-    LetterShift: 'С',
-    code: 'KeyC',
-  },
-  {
-    Letter: 'м',
-    LetterShift: 'М',
-    code: 'KeyV',
-  },
-  {
-    Letter: 'и',
-    LetterShift: 'И',
-    code: 'KeyB',
-  },
-  {
-    Letter: 'т',
-    LetterShift: 'Т',
-    code: 'KeyN',
-  },
-  {
-    Letter: 'ь',
-    LetterShift: 'Ь',
-    code: 'KeyM',
-  },
-  {
-    Letter: 'б',
-    LetterShift: 'Б',
-    code: 'Comma',
-  },
-  {
-    Letter: 'ю',
-    LetterShift: 'Ю',
-    code: 'Period',
-  },
-  {
-    Letter: '.',
-    LetterShift: ',',
-    code: 'Slash',
-  },
-  {
-    Letter: 'Shift',
-    LetterShift: 'Shift',
-    code: 'ShiftRight',
-  },
-  {
-    Letter: 'Ctrl',
-    LetterShift: 'Ctrl',
-    code: 'ControlLeft',
-  },
-  {
-    Letter: 'Alt',
-    LetterShift: 'Alt',
-    code: 'AltLeft',
-  },
-  {
-    Letter: ' ',
-    LetterShift: ' ',
-    code: 'Space',
-  },
-  {
-    Letter: 'Alt',
-    LetterShift: 'Alt',
-    code: 'AltRight',
-  },
-  {
-    Letter: 'Ctrl',
-    LetterShift: 'Ctrl',
-    code: 'ControlRight',
-  },
-  {
-    Letter: '&larr;',
-    LetterShift: '&larr;',
-    code: 'ArrowLeft',
-  },
-  {
-    Letter: '&uarr;',
-    LetterShift: '&uarr;',
-    code: 'ArrowUp',
-  },
-  {
-    Letter: '&darr;',
-    LetterShift: '&darr;',
-    code: 'ArrowDown',
-  },
-  {
-    Letter: '&rarr;',
-    LetterShift: '&rarr;',
-    code: 'ArrowRight',
-  },
-  {
-    Letter: '<div class="icon"></div>',
-    LetterShift: '<div class="icon"></div>',
-    code: 'LNG',
-  },
-];
