--- js/storageset.js
@@ -1,7 +1,0 @@
-export function set(name, value) {
-  window.localStorage.setItem(name, value);
-}
-
-export function get(name) {
-  return (window.localStorage.getItem(name) || null);
-}
